+++ litebird_sim/gaindrifts.py
@@ -5,7 +5,7 @@ from enum import IntEnum
 import numpy as np
 
 from .observations import Observation
-from .seeding import regenerate_or_check_detector_generators
+from .observation_utilities import for_each_observation
 
 
 class GainDriftType(IntEnum):
@@ -548,22 +548,13 @@ def apply_gaindrift_to_observations(
     if drift_params is None:
         drift_params = GainDriftParams()
 
-    if isinstance(observations, Observation):
-        obs_list = [observations]
-    elif isinstance(observations, list):
-        obs_list = observations
-    else:
-        raise TypeError(
-            "The parameter `observations` must be an `Observation` or a list of `Observation`."
-        )
-    dets_random = regenerate_or_check_detector_generators(
-        observations=obs_list,
+    for cur_obs, tod, dets_random in for_each_observation(
+        observations,
+        component,
         user_seed=user_seed,
         dets_random=dets_random,
-    )
-
-    for cur_obs in obs_list:
-        tod = getattr(cur_obs, component)
+        requires_rng=True,
+    ):
         sampling_freq_hz = cur_obs.sampling_rate_hz
         focalplane_attr = getattr(cur_obs, drift_params.focalplane_group)