+++ litebird_sim/mapmaking/destriper.py
@@ -16,7 +16,7 @@ from litebird_sim.coordinates import CoordinateSystem, coord_sys_to_healpix_stri
 from litebird_sim.hwp import HWP
 from litebird_sim.mpi import MPI_ENABLED, MPI_COMM_WORLD, MPI_COMM_GRID
 from litebird_sim.observations import Observation
-from litebird_sim.pointings_in_obs import (
+from litebird_sim.observation_utilities import (
     _get_hwp_angle,
     _normalize_observations_and_pointings,
 )